backend/pyproject.toml
[tool.poetry]
name = "NerdSwipe Backend"
version = "0.1.0"
description = "A FastAPI backend for NerdSwipe."
authors = ["Imlapps"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
pathlib = "^1.0.1"
langchain = "^0.1.11"
pydantic-settings = "^2.2.1"
unidecode = "^1.3.8"
langchain-openai = "^0.1.3"
pyoxigraph = "0.3.22"
supabase = "^2.8.1"
pyjwt = "^2.9.0"
python-multipart = "^0.0.12"
pydantic-extra-types = "^2.10.0"
pycountry = "^24.6.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
ruff = "^0.6.7"
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
mypy-extensions = "^1.0.0"
tee = "^0.0.3"
isort = "^5.13.2"
pytest-order = "^1.3.0"
pytest-asyncio = "^0.24.0"
asgi-lifespan = "^2.1.0"
trio = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"